+++ moat-cli/moat-cli-user/src/run_result.rs
@@ -4,11 +4,10 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-use rkyv::ser::serializers::AllocSerializer;
-use sha3::{Digest, Sha3_256};
 use std::fmt;
 use std::ops::Range;
 use zk_citadel::license::License;
+use zk_citadel_moat::MoatCoreUtils;
 
 pub struct SubmitRequestSummary {
     pub psk_lp: String,
@@ -70,7 +69,7 @@ impl fmt::Display for RunResult {
                         writeln!(
                             f,
                             "license: {} {}",
-                            RunResult::to_hash_hex(license),
+                            MoatCoreUtils::to_hash_hex(license),
                             if *is_owned { "owned" } else { "" }
                         )?;
                     }
@@ -83,7 +82,7 @@ impl fmt::Display for RunResult {
                         writeln!(
                             f,
                             "using license: {}",
-                            Self::blob_to_hash_hex(
+                            MoatCoreUtils::blob_to_hash_hex(
                                 summary.license_blob.as_slice()
                             )
                         )?;
@@ -96,7 +95,7 @@ impl fmt::Display for RunResult {
                         writeln!(
                             f,
                             "license {} used",
-                            Self::blob_to_hash_hex(
+                            MoatCoreUtils::blob_to_hash_hex(
                                 summary.license_blob.as_slice()
                             ),
                         )?;
@@ -126,41 +125,5 @@ impl fmt::Display for RunResult {
             }
             Empty => Ok(()),
         }
-    }
-}
-
-impl RunResult {
-    pub fn to_hash_hex<T>(object: &T) -> String
-    where
-        T: rkyv::Serialize<AllocSerializer<16386>>,
-    {
-        let blob = rkyv::to_bytes::<_, 16386>(object)
-            .expect("Serializing should be infallible")
-            .to_vec();
-        Self::blob_to_hash_hex(blob.as_slice())
-    }
-
-    pub fn blob_to_hash_hex(blob: &[u8]) -> String {
-        let mut hasher = Sha3_256::new();
-        hasher.update(blob);
-        let result = hasher.finalize();
-        hex::encode(result)
-    }
-
-    pub fn to_blob_hex<T>(object: &T) -> String
-    where
-        T: rkyv::Serialize<AllocSerializer<16386>>,
-    {
-        let blob = Self::to_blob(object);
-        hex::encode(blob)
-    }
-
-    pub fn to_blob<T>(object: &T) -> Vec<u8>
-    where
-        T: rkyv::Serialize<AllocSerializer<16386>>,
-    {
-        rkyv::to_bytes::<_, 16386>(object)
-            .expect("Serializing should be infallible")
-            .to_vec()
     }
 }
