+++ moat/src/utils.rs
@@ -38,12 +38,14 @@ pub struct SetupHolder {
 
 pub struct MoatCoreUtils {}
 
+const MAX_OBJECT_SIZE: usize = 16384;
+
 impl MoatCoreUtils {
     pub fn to_hash_hex<T>(object: &T) -> String
     where
-        T: rkyv::Serialize<AllocSerializer<16386>>,
+        T: rkyv::Serialize<AllocSerializer<MAX_OBJECT_SIZE>>,
     {
-        let blob = rkyv::to_bytes::<_, 16386>(object)
+        let blob = rkyv::to_bytes::<_, MAX_OBJECT_SIZE>(object)
             .expect("Serializing should be infallible")
             .to_vec();
         Self::blob_to_hash_hex(blob.as_slice())
@@ -54,6 +56,23 @@ impl MoatCoreUtils {
         hasher.update(blob);
         let result = hasher.finalize();
         hex::encode(result)
+    }
+
+    pub fn to_blob_hex<T>(object: &T) -> String
+    where
+        T: rkyv::Serialize<AllocSerializer<MAX_OBJECT_SIZE>>,
+    {
+        let blob = Self::to_blob(object);
+        hex::encode(blob)
+    }
+
+    pub fn to_blob<T>(object: &T) -> Vec<u8>
+    where
+        T: rkyv::Serialize<AllocSerializer<MAX_OBJECT_SIZE>>,
+    {
+        rkyv::to_bytes::<_, MAX_OBJECT_SIZE>(object)
+            .expect("Serializing should be infallible")
+            .to_vec()
     }
 
     pub async fn get_license_to_use(
